uncommented out Map map = MapControl.createMap(); //create and initialize new map        game.setMap(map); //same map in
added createScenes function to MapControl.java

+++ BigFishChallenge/src/byui/cit260/bigFishChallenge/model/Location.java
@@ -15,8 +15,8 @@ import java.util.Objects;
 public class Location implements Serializable {
     
     //class instance variables
-    private double row;
-    private double column;
+    private int row;
+    private int column;
     private boolean visited;
     private MainScene scene;
     private Actor actors;
@@ -50,7 +50,7 @@ public class Location implements Serializable {
         return row;
     }
 
-    public void setRow(double row) {
+    public void setRow(int row) {
         this.row = row;
     }
 
@@ -58,7 +58,7 @@ public class Location implements Serializable {
         return column;
     }
 
-    public void setColumn(double column) {
+    public void setColumn(int column) {
         this.column = column;
     }
